// src/server/model/iPaymentLine.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { TransactionInfo } from './transactionInfo';

export interface IPaymentLine { 
    transaction?: TransactionInfo;
    paymentModeName?: string;
    ledgerName?: string;
    ledgerId?: number;
    amount?: number;
    chequeDate?: Date;
    chequeNumber?: string;
    cardType?: string;
    cardNumber?: string;
    chequeBank?: string;
    paymentModeId?: number;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}