// src/server/model/iPaymentMode.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface IPaymentMode { 
    type?: IPaymentMode.TypeEnum;
    code?: string;
    ledgerName?: string;
    ledgerId?: number;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}
export namespace IPaymentMode {
    export type TypeEnum = 'Card' | 'Cheque' | 'WireTransfer' | 'Voucher' | 'MobileMoney' | 'Other' | 'Cash' | 'E_Commerce';
    export const TypeEnum = {
        Card: 'Card' as TypeEnum,
        Cheque: 'Cheque' as TypeEnum,
        WireTransfer: 'WireTransfer' as TypeEnum,
        Voucher: 'Voucher' as TypeEnum,
        MobileMoney: 'MobileMoney' as TypeEnum,
        Other: 'Other' as TypeEnum,
        Cash: 'Cash' as TypeEnum,
        ECommerce: 'E_Commerce' as TypeEnum
    };
}